requirements.txt
telegram-bot-api
pixivpy3
requests
pymongo